feat(dashboard-scene): add DashboardScene.addVariable() and delegate from ADD_VARIABLE handler
Centralises the v2-to-Scenes variable transformation inside a new
DashboardScene.addVariable(variable: VariableKind, position?) method.
The ADD_VARIABLE mutation API handler now delegates to it, keeping the
transformation logic in one place instead of duplicated across handler
and future UI call sites.
